requirements.txt
beautifulsoup4
langchain
langgraph
langchainhub
langchain-community
langchain_chroma
tavily-python
chromadb
python-dotenv
black
isort
pytest
grandalf
rich